GGW-203 Map - Add configuration to turn on and off sectors B
