+++ amp/TEMPLATE/ampTemplate/gisModule/dev/app/js/amp/app.js
@@ -37,35 +37,7 @@ if(state.loadPromise){
 	configureApp();
 }
 
-let cachedGisSettings = null;
 
-function getGisSettings() {
-	return new Promise((resolve, reject) => {
-		if (cachedGisSettings !== null) {
-			resolve(cachedGisSettings);
-			return;
-		}
-
-		fetch('/rest/amp/settings/gis')
-			.then(response => {
-				if (!response.ok) {
-					throw new Error('Network response was not ok');
-				}
-				return response.json();
-			})
-			.then(data => {
-				// Cache the GIS settings
-				cachedGisSettings = data;
-
-				// Resolve the promise with the JSON response
-				resolve(data);
-			})
-			.catch(error => {
-				// Reject the promise with the error
-				reject(error);
-			});
-	});
-}
 
 //configure to use saved language
 function setSavedLanguage(){
@@ -119,10 +91,5 @@ function configureApp() {
 
 }
 
-
-
-module.exports = {
-	getGisSettings,
-};
 
 module.exports = window.app = app;
